requirements.txt
pandas==2.2.2
requests==2.32.3
tenacity==8.5.0
polygon-api-client==1.0.0
python-dotenv
